valve_control/requirements.txt
RPi.GPIO>=0.7.0 